Signal handling implemented. The threads are now dying without leaving a mess.
Configuration structure was moved to global scope again so it is available
to signal handler. Access to the configuration structure from the
threads remains the same via internal pointer in thread data structure.

Controlling the execution (run/die) of threads is done through loop_state
variable which is available in thread data structure. Pointer to each
loop_state variable is available to the main thread and its signal handler
in threads_loop_state array.