Stub of etherpoke version 2.0.
This is not a functional version, it only prints out messages when events
are triggered. Event hooks are not yet supported.

Unlike the previous version, it no longer requires POSIX
threads to be available. Entire program was rewritten and simplified.
Asynchronicity was achieved through use of events (select(2)).

Configuration file uses different format and filter rules are now
written in BPF syntax.