Send event notifications via socket.
The program opens a TCP port (on user's demand using option -l on
program start) and waits for clients to connect. If event is triggered,
event hook (command) is sent via socket to all connected clients. By
default, socket notification is disabled for every filter, unless
configuration option 'notify_sock' is set to true.

In future versions, format of the socket notifications should be
recrafted as it is now impossible to distinguish filters with the same
event hooks.